[Keyless Common] Renamed config.rs -> circuit_config.rs
