+++ circuit/src/bigint.rs
@@ -2,7 +2,7 @@
 
 use crate::TestCircuitHandle;
 use aptos_keyless_common::input_processing::{
-    circuit_input_signals::CircuitInputSignals, config::CircuitConfig,
+    circuit_config::CircuitConfig, circuit_input_signals::CircuitInputSignals,
 };
 use aptos_logger::info;
 use rand::{thread_rng, Rng};
