+++ circuit/src/hash_to_field.rs
@@ -9,7 +9,7 @@ use aptos_crypto::{
     test_utils::random_bytes,
 };
 use aptos_keyless_common::input_processing::{
-    circuit_input_signals::CircuitInputSignals, config::CircuitConfig,
+    circuit_config::CircuitConfig, circuit_input_signals::CircuitInputSignals,
 };
 use ark_bn254::Fr;
 use ark_ff::Field;
