+++ circuit/src/jwt_field_parsing.rs
@@ -2,7 +2,7 @@
 
 use crate::{misc::calc_string_bodies, TestCircuitHandle};
 use aptos_keyless_common::input_processing::{
-    circuit_input_signals::CircuitInputSignals, config::CircuitConfig,
+    circuit_config::CircuitConfig, circuit_input_signals::CircuitInputSignals,
 };
 
 #[derive(Clone)]
