+++ circuit/src/sha.rs
@@ -2,7 +2,7 @@
 
 use crate::TestCircuitHandle;
 use aptos_keyless_common::input_processing::{
-    circuit_input_signals::CircuitInputSignals, config::CircuitConfig, sha,
+    circuit_config::CircuitConfig, circuit_input_signals::CircuitInputSignals, sha,
 };
 use rand_chacha::{
     rand_core::{RngCore as _, SeedableRng as _},
