+++ keyless-common/src/input_processing/mod.rs
@@ -1,8 +1,8 @@
 // Copyright (c) Aptos Foundation
 
 pub mod bits;
+pub mod circuit_config;
 pub mod circuit_input_signals;
-pub mod config;
 pub mod encoding;
 pub mod sha;
 pub mod witness_gen;
