+++ prover-service/src/request_handler/prover_state.rs
@@ -4,7 +4,7 @@ use crate::external_resources::jwk_fetcher::JWKCache;
 use crate::external_resources::prover_config::ProverServiceConfig;
 use crate::request_handler::deployment_information::DeploymentInformation;
 use aptos_crypto::ed25519::{Ed25519PrivateKey, Ed25519PublicKey};
-use aptos_keyless_common::input_processing::config::CircuitConfig;
+use aptos_keyless_common::input_processing::circuit_config::CircuitConfig;
 use rust_rapidsnark::FullProver;
 use std::sync::Arc;
 use tokio::sync::Mutex;
